pyproject.toml
[project]
name = "fx-quotes"
version = "0.1.0"
description = "Foreign exchange micro service in python/django."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "celery[redis]>=5.5.3",
  "coverage>=7.11.3",
  "django>=5.2.8",
  "django-environ>=0.12.0",
  "djangorestframework>=3.16.1",
  "ipdb>=0.13.13",
  "psycopg[binary]>=3.2.12",
  "redis>=5.2.1",
]
